+++ src/components/AdminContacts.vue
@@ -70,7 +70,8 @@
       },
 
       deleteRecord(id) {
-        this.$store.dispatch("contactsModule/deleteContact", id);
+        if (window.confirm('Tem certeza que deseja excluir este registro?'))
+          this.$store.dispatch("contactsModule/deleteContact", id);
       }
     },
