+++ src/components/AdminServices.vue
@@ -91,7 +91,8 @@ export default {
     },
 
     deleteRecord(id) {
-      this.$store.dispatch("servicesModule/deleteService", id);
+      if (window.confirm('Tem certeza que deseja excluir este registro?'))
+        this.$store.dispatch("servicesModule/deleteService", id);
     },
   },
